move S3 helpers to S3 file
